Feature/TF-211 lineup integration

+++ components/Includes/Banner/StickyAds.js
@@ -38,7 +38,7 @@ class StickyAds extends React.Component {
   }
 
   componentWillUnmount() {
-    googletag.destroySlots()
+    if (googletag.destroySlots) googletag.destroySlots()
   }
 
   render() {
